+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Remove absolute path for boost library ([#125](https://github.com/Simple-Robotics/proxsuite-nlp/pull/125))
+- Remove absolute path for boost library in generated pkg-config file ([#125](https://github.com/Simple-Robotics/proxsuite-nlp/pull/125))
 
 ### Added